backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai==0.3.0
openai==1.58.1
websockets==14.1
mangum==0.19.0
PyPDF2==3.0.1
psutil==5.9.1
httpx==0.27.0
